https://bugs.eclipse.org/bugs/show_bug.cgi?id=372417: "Advanced mode pattern"
new content and label provider usage for super type fields of wizards (make use of the lazy loading). this way the proposals are only loaded when modified by the user. this helps to prevent calculating unneeded type hierarchies (performance). especially in big workspaces this improves the open time of wizards when opened the first time.