Add archetype for a Scout application with JavaScript UI
Adds an archetype that uses ScoutJS for the UI. The communication with
the server which exposes a REST API based on JSON objects. The
persistence layer uses jooq to access the database.
Furthermore the Scout new project wizard allows to choose between the
Scout Classic archetype using Java for the UI and the ScoutJS archetype
using JavaScript for the UI.