Move Operations to SDK Core module
Operations should be reusable independent of the environment (IDE) in
which they are running. Therefore they have been moved to SDK Core
module. All requirements to the environment have been extracted to a
single interface which allows to implement it by IDE.
Furthermore the source builders have been replaced by source generators
using a fluent API using Java 8 features.
Finally a Scout SDK plugin for IntelliJ IDEA is added which makes use of
the newly available operations in the core.