IFuture result computation should be lazy
The result computation of IFutures should only be performed if the client requests the result. If it just waits for the computation to complete, the result must not be computed. Therefore the result is always of type Supplier so that the computation of the supplier value is only performed on request and only once.