Temp directory should be deleted after the test
If mvn clean verify is executed the WarMojo runs as well. During the
execution the method
org.apache.maven.plugins.war.WarMojo#isProjectUsingAtLeastServlet30() is
executed which creates an URLClassLoader without properly closing it.
This leaks open file handles which prevent the temp directory to be
deleted after the test.
Workaround: only execute mvn clean test. The phases between test and
verify are not of interest anyway and the WarMojo will not be executed.