Archetypes: simplify launch files
The existing node js script has some problems:
1. If node is node installed, a very cryptic message is printed which
does not help at all.
2. On OS X, even if node is installed, it does not work because
the PATH variable is not the same as on the terminal -> the same
cryptic message appears.
3. When starting the script in debug mode it won't work with newer
node installations because eclipse does not support it -> another
cryptic (a little bit less cryptic) message appears.
4. Users don't understand that the scripts only starts 2 command line
tools: npm install and npm run build:dev:watch.

-> Solution is to replace the Node scripts with shells scripts.
It is now possible to print a better message if node is not installed.
PATH is set by default to the common install directory on os x.
When running in debug mode a clear message appears.
It is hopefully easier to understand for more people.

Further changes:
1. The watcher is always started. This has the benefit that
CSS and JS bundles will always be rebuilt if necessary. And, maybe
even more important for people not writing JS or CSS, running the launch
group (Launch all) again won't rerun the js build script because it is
already running due to the watcher.
2. The ui server is not started unless the js build prints "Built at:"
which indicates the web pack build has finished -> it won't start if
npm cannot be found or the webpack build fails.