Type annotations must use the outer import collector
An annotation on a type is not a member of the type. Instead the
annotation exists on the same level as the type. Therefore the import
collector context is the same as the declaring element of the type not
the collector of the members of the type.
Example:
A primary type with an annotation pointing to a member type of the
primary type. Because the annotation exists on the same level as the
primary type (not within the primary type) the access to the member type
 must either be qualified or an import must be created.
Therefore call the creation of the annotation with the current import
collector instead of using the EnclosingTypeScopedImportCollector which
is used for members of the primary type.