Async write operations should run in same transaction as the caller
Until now in IntelliJ an async compilation unit write operation always
used its own transaction for each write. This is different than in
Eclipse and makes no sense from a parent context perspective because the
 result of the write operation should only be committed when all async
 operations are successful.
Therefore clarify the actual contract in the JavaDoc of the async
methods and ensure the forked async write operation runs in the parent
transaction.