Move declaringType() method from IType to IMember
As all IMember instances may exist in a type it makes sense to have the
declaringType() method on the same level.
As IMethod and IField always have a declaring type, a convenience method
requireDeclaringType() has been added which directly returns the type
instead of returning an Optional as declaringType() does.