Find translation stores in Scout JS applications
In Scout JS applications (like the hellojs app) all translations from
the backend are transferred to the browser at the application startup
and are available to be used from JavaScript code.
This is the case even though the module with the JavaScript code has no
dependency to the backend (neither using Maven nor using Node).

To make all the texts from the backend available to the JavaScript code
the following naming convention is used: Such a JavaScript module ends
with ".ui" and its backend entry point with ".app".
So for all modules ending with ".ui" the translations of a sibling
module with suffix ".app" is included as well.