Translation completion: only insert trailing "}" if necessary
For Scout JS text key translation code completion the trailing "}"
should only be inserted if not already present.
This is not possible with default insertion. Therefore provide a custom
insertion handler.
To provide the language dependent new psi element which should be
inserted a TranslationLanguageSpec is introduced. It holds all language
specific translation logic. For now three implementations are provided:
Java, Js, Html.