Fix automatic override annotation creation with type arguments
If a super method is declared using type parameter data types, it should
be correctly detected based on the context of the child class.
E.g. if a super class specifies a method 'm(T)' and a subclass
implements this method like 'm(String)' because T is bound to String, it
 should even though be detected as super method and therefore an
 override annotation should be added.