Correctly handle types in IJavaEnvironment#reload
The old implementation of internalFindNewElement simply selected a new
type with the same fully qualified name. While this is correct for type
declarations it is wrong for reference bindings.
The new implementation tries to select the exactly same location again
as the type was created originally.
Furthermore the AbstractJavaEnvironment#reload implementation was
hardened to better handle SPI<->API collisions in case multiple SPIs
resolve to the same API. Such detached APIs are now retained and will
continue to get updates in future reload operations.