Use the users format locale as default translation language
This better matches a language spoken by the user.
It uses the FORMAT category instead of the DISPLAY category because the
DISPLAY might correlate to the text language of the operating system
which is not necessarily the language spoken by the user. E.g. some
developers use English systems even though their mother tongue is
different. The FORMAT category better matches the locale of the user.