Change behavior of NLS editor
In former versions the editor only showed texts for a key from the text
provider service with the highest order. Overridden texts have been
hidden even if the highest service has no own text for a language.
This was very confusing as at runtime the text could be found
successfully.

Therefore now the table for each key shows all available texts from all
stores containing that key. It respects the order of the text services
to correctly override texts as it would be at runtime.

This has the consequence that one row in the table may belong to several
text services now. This has some impact on the operations:
1. Deleting a text is only possible if all stores of that row are
   editable.
2. Changing the key is only possible if all stores of a row are editable