Add TierTree and ITier
Add a tree that represents the hierarchy of all tiers. This tree can
handle ITier's, therefore ScoutTier now implements ITier. The TierTree
provides the following methods
 - addDependency: add a direct dependency between two tiers
 - hasDependency: check if a tier has a direct or transitive dependency
   on another given tier
 - isAvailable: check if a tier equals another given tier or has a
   dependency on it
 - path: get the tier path from one tier to another
 - tierOf: as a tier is a Predicate<IJavaElement>, this method searches
   from the leaves to the root node for a matching ITier
The SourceFolderHelper now uses the TierTree to calculate all dependent
and dependency sourceFolders.
All hard coded dependency hierarchy between the ScoutTier was removed
and transferred to the TierTree in order to make the dependency
hierarchy easily extendable.