JS model completion: add class references for objectType
Since 23.1 scout.create works with class references instead of string
literals. Add support for this feature to the JS model completion.
The new scout.create-feature also comes with a new way of registering
the namespace of a module. The JS model completion now uses this new
namespace-pattern to detect whether a module works with class references
or not. If a class reference is added, the namespace will be omitted and
an import is added if necessary.

327445