JS Code completion: improve extended prefix
If one completes an enum property value and selects a value like
"Form.DisplayHint.VIEW" in an empty space it is inserted completely. But
if the same value is selected with a prefix like "Form.DisplayH" only
the value "DisplayHint.VIEW" needs to be inserted as IntelliJ replaces
everything from the last "." to the current position of the cursor (in
this case: "DisplayH").
Remove unnecessary insertHandler for quotes, but manipulate the
lookupString accordingly.