IdeaPackageJson: Handle possible IOException
The IdeaPackageJson is created with a module directory, and it is
ensured that this directory contains a package.json file at instance
creation time.
The content of this package.json is only read and parsed on first use.
Since the instance creation the package.json file might have been
removed. In this case parsing the file content should not throw an
IOException but instead this module should be ignored (as it is no
longer present anyway).