ScoutJsModel: decide whether TS or JS by containing file
Decide whether to parse TS or JS objects/enums by the containing file of
the element and not by the main file of the containing module.
There are modules containing valid JS and TS objects/enums. Parsing only
the TS or only the JS objects/enums will result in an incomplete list of
objects/enums.

356286