Downgrade ecj from 3.35 to 3.34
This is necessary because ecj has issues resolving constant Annotation
values pointing e.g. to a String in a super class of a nested Class.

Example where XZY is defined in SuperIfc.Names as String constant:
@Annot(myVal = Test.Names.XYZ)
interface Test {
  interface Names extends SuperIfc.Names {
  }
}
In this situation the compiler can no longer resolve the constant value
of Annotation myVal.

This seems to work in ecj 3.34.