Fix creating new projects in Eclipse
The workspace resolution in a maven build in eclipse fails iff there are
no modules because the file
WS_DIR\.metadata\.plugins\org.eclipse.m2e.core\workspacestate.properties
does not exist.
Therefore, a marker property is added to all maven builds creating a new
project and the workspace resolution is turned off if this marker
property is set.